pyproject.toml
[tool.poetry]
name = "get-wb-feedback"
version = "0.1.0"
description = ""
authors = ["Polina Dikaya <maksimovapi@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-api-python-client = "^2.114.0"
python-dotenv = "^1.0.0"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
